Applying feedback and moving to own package.
