+++ jetbrains-core/resources/META-INF/ext-terminal.xml
@@ -3,6 +3,6 @@
 
 <idea-plugin>
     <actions>
-        <action id="aws.toolkit.startTerminal" class="software.aws.toolkits.jetbrains.core.StartAwsTerminal"/>
+        <action id="aws.toolkit.startTerminal" class="software.aws.toolkits.jetbrains.core.terminal.StartAwsTerminal"/>
     </actions>
 </idea-plugin>
